# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel Management API",
        "description": "API REST para gerenciamento de pedidos de viagem corporativa. Autenticação via JWT Bearer token.",
        "contact": {
            "email": "marinairis.dev@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/travel-requests": {
            "get": {
                "tags": [
                    "Travel Requests"
                ],
                "summary": "Listar pedidos de viagem",
                "description": "Admin/Manager veem todos os pedidos. Solicitante vê apenas os próprios.",
                "operationId": "39bccb080fd6a513157990217625fb89",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "requested",
                                "approved",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "destination",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "São Paulo"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-01-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-12-31"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de pedidos"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Travel Requests"
                ],
                "summary": "Criar novo pedido de viagem",
                "operationId": "e2e47f7ce74debd6cdea7a77644b707e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "requester_name",
                                    "destination",
                                    "departure_date",
                                    "return_date"
                                ],
                                "properties": {
                                    "requester_name": {
                                        "type": "string",
                                        "example": "Maria Silva"
                                    },
                                    "destination": {
                                        "type": "string",
                                        "example": "São Paulo, SP"
                                    },
                                    "departure_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-08-01"
                                    },
                                    "return_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-08-05"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Reunião com cliente"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido criado com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/travel-requests/{id}": {
            "get": {
                "tags": [
                    "Travel Requests"
                ],
                "summary": "Consultar pedido de viagem por ID",
                "operationId": "f0b25709c89b5765768c15f949e84eb7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do pedido"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Travel Requests"
                ],
                "summary": "Atualizar dados de um pedido de viagem",
                "operationId": "dbfb18f7e2c2a806be8b171399326a13",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "requester_name": {
                                        "type": "string"
                                    },
                                    "destination": {
                                        "type": "string"
                                    },
                                    "departure_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "return_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pedido atualizado"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão ou pedido já aprovado"
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Travel Requests"
                ],
                "summary": "Excluir pedido de viagem (Gestor ou Admin)",
                "description": "Pedidos aprovados não podem ser excluídos.",
                "operationId": "9a592eb222bda25c16fb281f09f6ab72",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido excluído"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão ou pedido aprovado"
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/travel-requests/{id}/status": {
            "patch": {
                "tags": [
                    "Travel Requests"
                ],
                "summary": "Atualizar status de um pedido (aprovado/cancelado)",
                "description": "Apenas Gestor ou Admin. O usuário que criou o pedido não pode alterar seu próprio status.",
                "operationId": "a78288b20a2794c144eb26a1ffa83a18",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "approved",
                                            "cancelled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status atualizado com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    },
                    "422": {
                        "description": "Status inválido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/travel-requests/{id}/cancel": {
            "patch": {
                "tags": [
                    "Travel Requests"
                ],
                "summary": "Cancelar pedido de viagem",
                "description": "Pedidos aprovados só podem ser cancelados se a data de partida ainda não passou.",
                "operationId": "117fe900711302da9d6a2cb29c6c7638",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido cancelado com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    },
                    "422": {
                        "description": "Pedido já cancelado ou data de partida passou"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}